services/feedback-service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
pydantic
pydantic-settings
python-dotenv
pytest
pytest-asyncio
httpx
aiosqlite
